basic rich release description
