+++ js/hubs/styles/theme/lightThemeOptions.js
@@ -205,6 +205,9 @@ const lightThemeOptions = {
             opacity: "50% !important",
           },
         },
+        'p, h1, h2, h3, h4, figure, blockquote, dl, dd': {
+          margin: 0
+        },
         ".formField": {
           marginBottom: "8px !important",
           width: "100%",
